return test_build but skip without xelatex
