+++ tests/common.py
@@ -1,3 +1,6 @@
+import subprocess
+from typing import List
+
 DEMO_MAPPING = {
     "source/main.ptx": ["my-demo-book"],
     "source/frontmatter.ptx": [
@@ -24,3 +27,14 @@ DEMO_MAPPING = {
     ],
     "source/backmatter.ptx": ["backmatter"],
 }
+
+# Return True if the given binary is installed and exits with a return code of 0; otherwise, return False. This provides an easy way to check that a given binary is installed.
+def check_installed(
+    # The command to run to check that a given binary is installed; for example, `["python", "--version"]` would check that Python is installed.
+    subprocess_args: List[str],
+) -> bool:
+    try:
+        subprocess.run(subprocess_args, check=True)
+    except Exception:
+        return False
+    return True
