+++ tests/test_cli.py
@@ -11,14 +11,17 @@ import requests
 import pretext
 from pretext import constants
 from typing import cast, Generator
+import pytest
 from pytest_console_scripts import ScriptRunner
-from .common import DEMO_MAPPING
+from .common import DEMO_MAPPING, check_installed
 
 EXAMPLES_DIR = Path(__file__).parent / "examples"
 
 PTX_CMD = cast(str, shutil.which("pretext"))
 assert PTX_CMD is not None
 PY_CMD = sys.executable
+
+HAS_XELATEX = check_installed(["xelatex", "--version"])
 
 
 @contextmanager
@@ -69,17 +72,20 @@ def test_devscript(script_runner: ScriptRunner) -> None:
     assert "PreTeXt utility script" in result.stdout
 
 
+@pytest.mark.skipif(
+    not HAS_XELATEX,
+    reason="Skipped since xelatex isn't found.",
+)
 def test_build(tmp_path: Path, script_runner: ScriptRunner) -> None:
     path_with_spaces = "test path with spaces"
     project_path = tmp_path / path_with_spaces
     assert script_runner.run(
-        [PTX_CMD, "-v", "debug", "new", "demo", "-d", path_with_spaces],
-        cwd=tmp_path,
+        [PTX_CMD, "-v", "debug", "new", "demo", "-d", path_with_spaces], cwd=tmp_path
     ).success
 
     # Do a subset build before the main build, to check that not everything is built on the subset.
     assert script_runner.run(
-        [PTX_CMD, "-v", "debug", "build", "web", "-x", "ch-first-without-spaces", "-q"],
+        [PTX_CMD, "-v", "debug", "build", "web", "-x", "ch-first-without-spaces"],
         cwd=project_path,
     ).success
     assert (project_path / "output" / "web").exists()
@@ -87,24 +93,21 @@ def test_build(tmp_path: Path, script_runner: ScriptRunner) -> None:
     assert (project_path / "output" / "web").stat().st_mode % 0o1000 >= 0o755
     assert not (project_path / "output" / "web" / "ch-empty.html").exists()
     assert (project_path / "output" / "web" / "ch-first-without-spaces.html").exists()
-    # TODO return this behavior to test suite
-    # needs to be dummied out for now since will result in a `log.error` if LaTeX
-    # is not available
-    # # Also do a subset without assets
-    # assert script_runner.run(
-    #     [PTX_CMD, "-v", "debug", "build", "web", "-x", "sec-latex-image", "-q"],
-    #     cwd=project_path,
-    # ).success
-    # assert not (project_path / "generated-assets" / "latex-image").exists()
-    # assert script_runner.run(
-    #     [PTX_CMD, "-v", "debug", "build", "web", "-x", "sec-latex-image"],
-    #     cwd=project_path,
-    # ).success
-    # assert (project_path / "generated-assets" / "latex-image").exists()
+    # Also do a subset without assets
+    assert script_runner.run(
+        [PTX_CMD, "-v", "debug", "build", "web", "-x", "sec-latex-image", "-q"],
+        cwd=project_path,
+    ).success
+    assert not (project_path / "generated-assets" / "latex-image").exists()
+    assert script_runner.run(
+        [PTX_CMD, "-v", "debug", "build", "web", "-x", "sec-latex-image"],
+        cwd=project_path,
+    ).success
+    assert (project_path / "generated-assets" / "latex-image").exists()
 
     # Do a full build.
     assert script_runner.run(
-        [PTX_CMD, "-v", "debug", "build", "web", "-q"], cwd=project_path
+        [PTX_CMD, "-v", "debug", "build", "web"], cwd=project_path
     ).success
     web_path = project_path / "output" / "web"
     assert web_path.exists()
