+++ tests/test_project.py
@@ -3,8 +3,6 @@ import json
 from pathlib import Path
 import requests
 import shutil
-import subprocess
-from typing import List
 
 import pydantic
 import pytest
@@ -13,25 +11,11 @@ from pretext import project as pr
 from pretext import templates
 from pretext import utils
 
-from .common import DEMO_MAPPING
+from .common import DEMO_MAPPING, check_installed
 
 
 EXAMPLES_DIR = Path(__file__).parent / "examples"
 TEMPLATES_DIR = Path(__file__).parent.parent / "templates"
-
-
-# Return True if the given binary is installed and exits with a return code of 0; otherwise, return False. This provides an easy way to check that a given binary is installed.
-def check_installed(
-    # The command to run to check that a given binary is installed; for example, `["python", "--version"]` would check that Python is installed.
-    subprocess_args: List[str],
-) -> bool:
-    try:
-        subprocess.run(subprocess_args, check=True)
-    except Exception:
-        return False
-    return True
-
-
 HAS_XELATEX = check_installed(["xelatex", "--version"])
